add weekHoursSpan param to validator analytics call and rename timeHelpers to analyticsTimeHelpers
